Extract array matching logic
In preparation to add array diffs to the failure messages of `match` the code to generate these diffs is extracted from `contain_exactly`.
